feat: InputValidator 클래스 추가
